delete AnalysisManager, all API functionality lives in Spotify.swift
